# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"MLB Highlights Project","version":"0.1.0"},"paths":{"/preferences/user/":{"post":{"tags":["preferences"],"summary":"Create User","operationId":"create_user_preferences_user__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preferences/set/":{"post":{"tags":["preferences"],"summary":"Set Preferences","operationId":"set_preferences_preferences_set__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreferenceCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mlb/schedule/":{"get":{"tags":["MLB Highlights"],"summary":"Get Schedule","description":"Get the MLB schedule for a given season and game type.","operationId":"get_schedule_mlb_schedule__get","parameters":[{"name":"season","in":"query","required":true,"schema":{"type":"integer","title":"Season"}},{"name":"game_type","in":"query","required":false,"schema":{"type":"string","default":"R","title":"Game Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mlb/team/{team_id}/roster/{season}":{"get":{"tags":["MLB Highlights"],"summary":"Get Team Roster","description":"Get the roster for a specific team and season.","operationId":"get_team_roster_mlb_team__team_id__roster__season__get","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}},{"name":"season","in":"path","required":true,"schema":{"type":"integer","title":"Season"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mlb/team/{team_id}/info/":{"get":{"tags":["MLB Highlights"],"summary":"Get Team Info","description":"Get detailed information about a specific team.","operationId":"get_team_info_mlb_team__team_id__info__get","parameters":[{"name":"team_id","in":"path","required":true,"schema":{"type":"integer","title":"Team Id"}},{"name":"season","in":"query","required":false,"schema":{"type":"integer","title":"Season"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mlb/player/{player_id}/info/":{"get":{"tags":["MLB Highlights"],"summary":"Get Player Info","description":"Get detailed information about a specific player.","operationId":"get_player_info_mlb_player__player_id__info__get","parameters":[{"name":"player_id","in":"path","required":true,"schema":{"type":"integer","title":"Player Id"}},{"name":"season","in":"query","required":false,"schema":{"type":"integer","title":"Season"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mlb/game/{game_pk}/live/":{"get":{"tags":["MLB Highlights"],"summary":"Get Live Game Data","description":"Get live game data for a specific game.","operationId":"get_live_game_data_mlb_game__game_pk__live__get","parameters":[{"name":"game_pk","in":"path","required":true,"schema":{"type":"integer","title":"Game Pk"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mlb/game/{game_pk}/timestamps/":{"get":{"tags":["MLB Highlights"],"summary":"Get Game Timestamps","description":"Get the list of update timestamps for a specific game.","operationId":"get_game_timestamps_mlb_game__game_pk__timestamps__get","parameters":[{"name":"game_pk","in":"path","required":true,"schema":{"type":"integer","title":"Game Pk"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mlb/game/{game_pk}/timecode/":{"get":{"tags":["MLB Highlights"],"summary":"Get Game At Timecode","description":"Get game data at a specific timecode.","operationId":"get_game_at_timecode_mlb_game__game_pk__timecode__get","parameters":[{"name":"game_pk","in":"path","required":true,"schema":{"type":"integer","title":"Game Pk"}},{"name":"timecode","in":"query","required":true,"schema":{"type":"string","title":"Timecode"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PreferenceCreate":{"properties":{"user_id":{"type":"integer","title":"User Id"},"team_ids":{"items":{"type":"integer"},"type":"array","title":"Team Ids"},"player_ids":{"items":{"type":"integer"},"type":"array","title":"Player Ids"}},"type":"object","required":["user_id","team_ids","player_ids"],"title":"PreferenceCreate"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language","default":"en"}},"type":"object","required":["username","email","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}